Split implementation in header an source files.
